IntRate, Rate, ReviewsCount

+++ src/main/java/com/heroku/demo/event/EventServiceImpl.java
@@ -143,8 +143,8 @@ public class EventServiceImpl implements EventService {
                     case 1: return Long.compare(o2.getId(), o1.getId());
                     case 2: return Integer.compare(o1.getPrice(), o2.getPrice());
                     case 3: return Integer.compare(o2.getPrice(), o1.getPrice());
-                    case 4: return Integer.compare(o1.getRate(), o2.getRate());
-                    case 5: return Integer.compare(o2.getRate(), o1.getRate());
+                    case 4: return Float.compare(o1.getRate(), o2.getRate());
+                    case 5: return Float.compare(o2.getRate(), o1.getRate());
                     case 6: return Integer.compare(o1.getType(), o2.getType());
                     case 7: return Integer.compare(o2.getType(), o1.getType());
                     case 8: return Integer.compare(o1.getDuration(), o2.getDuration());
